Initial D APIs for cloning, fetching, looking up objects, …
Sorry for the big messy commit and the half-baked state this is in,
(e.g. no tests) but I was in a hurry to get something working using
dlibgit.

The files I added don't include any of the original comments, as
they might be copyrightable, and would thus render the code GPL.